Add files via upload
Churn_Modelling
        Introduction
This is a predictive model for a customer in a bank predicting whether a customer is excited with the bank or not, whether the customer is going to leave the bank's services or not based on the behavior with their products.

     Variables used.
RowNumber: a unique identifier for each record.
CustomerId: a unique identifier for each customer.
Surname: the surname of the customer.
CreditScore: credit score of customer.
Geography: country of the customer.
Age: age of the customer.
Gender: whether a male or female.
Tenure: period which the customer has been in the bank.
Balance: customers account.
NumOfProducts: number of bank products the customer is using.
HasCrCard: whether the customer has a credit card or not.
IsActiveMember: whether the customer is active or not.
Salary:  Estimated salary of the customer.
Exited: indicates whether the customer is excited with the bank or not.

          Steps Followed
load the customer churn data into the panda's data frame. 
perform explanatory data analysis by checking the head of the data, the shape of the data, for missing values, and plotting the distribution of numerical and categorical variables.
prepare the data for the model by splitting it into training and test sets,one-hot encoding categorical variables, and fitting a logistic regression model
predict customer churn using the logistic regression model and evaluate its performance using the classification report

                     Explanation
The code first loads the data into a panda's data frame and performs EDA operations such as statistical summary, distribution of target columns, and checking missing values.
 
Next, it then encodes the categorical features and concatenates the encoded features with the original data.

Lastly, it splits the data into training and testing sets and fits a Logistic Regression model on the training data. It then predicts the target on the test data and evaluates the model using the confusion and classification report.

               Conclusion
With Machine Learning and deep analysis for customer churn prediction, the bank company will be able to increase customer retention rate and protect the future from churning the users excited with the bank.